fixed warnings

+++ src/solver.hpp
@@ -48,7 +48,7 @@ namespace evo
     protected:
         Model model;                             // instance of the model to be solved
         std::vector<int> n_obs;                  // number of observations for each trait
-        int n_trait;                             // number of traits
+        size_t n_trait;                             // number of traits
         std::vector<std::vector<int>> n_lev;     // num of effects' levels for each trait
         std::vector<matrix<float>> z_dat;        // combined incidence matrix for each trait.
         std::vector<matrix<float>> y;            // Observations for each trait.
